Replace simulation cases with filament extension example
- Remove previous simulation examples
- Add filamentExt.c as the main example case for the FIBER project
- Focus on flow-induced breakup scenarios
