+++ README.md
@@ -1,12 +1,11 @@
-# ElastoFlow: 2D/3D Viscoelastic Fluid Simulation Framework
+FIBER: Flow-Induced Breakup of Elastic Rheology
 
-[![DOI](https://zenodo.org/badge/874384071.svg)](https://doi.org/10.5281/zenodo.14011692)
-[![Website Status](https://img.shields.io/website?url=https%3A%2F%2Fcomphy-lab.github.io%2FViscoelastic3D%2F&style=flat-square&logo=github&label=Website)](https://comphy-lab.github.io/Viscoelastic3D/)
-[![GitHub](https://img.shields.io/badge/GitHub-Repository-blue?style=flat-square&logo=github)](https://github.com/comphy-lab/Viscoelastic3D)
+[![Website Status](https://img.shields.io/website?url=https%3A%2F%2Fcomphy-lab.github.io%2Ffiber%2F&style=flat-square&logo=github&label=Website)](https://comphy-lab.github.io/fiber/)
+[![GitHub](https://img.shields.io/badge/GitHub-Repository-blue?style=flat-square&logo=github)](https://github.com/comphy-lab/fiber)
 
 ---
 
-🚀 **ElastoFlow** is a state-of-the-art, open-source framework for simulating viscoelastic fluid flows in 2D and 3D, built as an extension to the [Basilisk C](http://basilisk.fr/) CFD library. It implements the log-conformation method for robust, high-Weissenberg number simulations, with a focus on clarity, extensibility, and scientific rigor.
+🚀 **Fiber** is a state-of-the-art, open-source framework for simulating viscoelastic fluid flows in 2D and 3D, built as an extension to the [Basilisk C](http://basilisk.fr/) CFD library. It implements the log-conformation method for robust, high-Weissenberg number simulations, with a focus on clarity, extensibility, and scientific rigor.
 
 ## ✨ Key Features
 
@@ -17,7 +16,7 @@
 - **Error Handling**: Negative eigenvalue checks, eigenvalue clamping, and detailed diagnostics
 - **Performance**: Simplified acceleration term calculations and optimized tensor operations
 - **Documentation**: Extensive inline documentation, mathematical background, and verification notes
-- **Compatibility**: GPLv3 license, fully compatible with Basilisk and previous ElastoFlow versions
+- **Compatibility**: GPLv3 license, fully compatible with Basilisk and previous versions
 
 ## 🐛 Bug Fixes (v2.5/v2.6)
 - Corrected matrix algebra in 3D
