+++ docs/index.html
@@ -167,7 +167,7 @@
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <meta name="base-url" content="..">
-  <title>ElastoFlow: 2D/3D Viscoelastic Fluid Simulation Framework | ElastoFlow: 2D/3D Viscoelastic Fluid Simulation Framework</title>
+  <title>FIBER: Flow-Induced Breakup of Elastic Rheology | FIBER: Flow-Induced Breakup of Elastic Rheology</title>
   
   <!-- Critical CSS for fastest paint -->
   <style>
@@ -332,7 +332,7 @@
   </script>
 </head>
 <body id="top">
-<script>window.repoName = "Viscoelastic3D";</script>
+<script>window.repoName = "fiber";</script>
 
 
     <!-- Preloader with theme variables applied -->
@@ -352,9 +352,9 @@
                 <a class="logo" href="/">
                     <img src="./assets/logos/CoMPhy-Lab-no-name.png" alt="CoMPhy Lab">
                 </a>
-                <a class="documentation-button" href="/Viscoelastic3D">
+                <a class="documentation-button" href="/fiber">
                     <i class="fa-solid fa-book-open"></i>
-                    <span>Viscoelastic3D</span>
+                    <span>FIBER</span>
                 </a>
             </div>
             <a class="s-header__menu-toggle" href="#0">
@@ -415,20 +415,20 @@
                                 
                 
                 <h1
-                id="elastoflow-2d3d-viscoelastic-fluid-simulation-framework">ElastoFlow:
-                2D/3D Viscoelastic Fluid Simulation Framework</h1>
+                id="fiber-flow-induced-breakup-of-elastic-rheology">FIBER:
+                Flow-Induced Breakup of Elastic Rheology</h1>
                 <p><a
                 href="https://doi.org/10.5281/zenodo.14011692"><img
                 src="https://zenodo.org/badge/874384071.svg"
                 alt="DOI" /></a> <a
-                href="https://comphy-lab.github.io/Viscoelastic3D/"><img
-                src="https://img.shields.io/website?url=https%3A%2F%2Fcomphy-lab.github.io%2FViscoelastic3D%2F&amp;style=flat-square&amp;logo=github&amp;label=Website"
+                href="https://comphy-lab.github.io/fiber/"><img
+                src="https://img.shields.io/website?url=https%3A%2F%2Fcomphy-lab.github.io%2Ffiber%2F&amp;style=flat-square&amp;logo=github&amp;label=Website"
                 alt="Website Status" /></a> <a
-                href="https://github.com/comphy-lab/Viscoelastic3D"><img
+                href="https://github.com/comphy-lab/fiber"><img
                 src="https://img.shields.io/badge/GitHub-Repository-blue?style=flat-square&amp;logo=github"
                 alt="GitHub" /></a></p>
                 <hr />
-                <p>🚀 <strong>ElastoFlow</strong> is a state-of-the-art,
+                <p>🚀 <strong>FIBER</strong> is a state-of-the-art,
                 open-source framework for simulating viscoelastic fluid
                 flows in 2D and 3D, built as an extension to the <a
                 href="http://basilisk.fr/">Basilisk C</a> CFD library.
@@ -462,7 +462,7 @@
                 documentation, mathematical background, and verification
                 notes</li>
                 <li><strong>Compatibility</strong>: GPLv3 license, fully
-                compatible with Basilisk and previous ElastoFlow
+                compatible with Basilisk and previous
                 versions</li>
                 </ul>
                 <h2 id="bug-fixes-v2.5v2.6">🐛 Bug Fixes
@@ -720,7 +720,7 @@
               <a href="https://bsky.app/profile/comphy-lab.org" target="_blank" aria-label="Bluesky Profile">
                 <i class="fa-brands fa-bluesky" style="font-size: 2.5em; color: white;"></i>
               </a>
-              <a href="https://github.com/comphy-lab/Viscoelastic3D" class="edit-link" aria-label="Edit this page on GitHub">
+              <a href="https://github.com/comphy-lab/fiber" class="edit-link" aria-label="Edit this page on GitHub">
                 <i class="fa-brands fa-github"></i> Edit this page
               </a>
             </div>
